+++ lib/masamune_flutter.dart
@@ -29,6 +29,7 @@ part 'core/defaultpath.dart';
 part 'core/bootstrap.dart';
 
 part 'component/uimaterialapp.dart';
+part 'component/widgetdata.dart';
 part 'component/uipage.dart';
 part 'component/extensions.dart';
 part 'component/uiwidget.dart';