+++ lib/widget/dialog/uiconfirm.dart
@@ -20,12 +20,12 @@ class UIConfirm {
   /// [dialogSubmitActionPath]: The path of action when the submit button of the dialog is pressed.
   /// [dialogCancelTextPath]: Dialog cancel button text path.
   /// [dialogCancelActionPath]: The path of action when the cancel button of the dialog is pressed.
-  /// [defaultSubmitText]: Default submit button text.
-  /// [defaultCancelText]: Default cancel button text.
-  /// [defaultSubmitAction]: Default submit button action.
-  /// [defaultCancelAction]: Default cancel button action.
-  /// [dialogTitle]: Default title.
-  /// [dialogText]: Default text.
+  /// [submitText]: Default submit button text.
+  /// [cancelText]: Default cancel button text.
+  /// [submitAction]: Default submit button action.
+  /// [cancelAction]: Default cancel button action.
+  /// [title]: Default title.
+  /// [text]: Default text.
   /// [popOnPress]: True if the dialog should be closed together when the button is pressed.
   /// [willShowRepetition]: True if the dialog will continue to be displayed unless you press the regular close button.
   static Future show(BuildContext context,
@@ -35,18 +35,18 @@ class UIConfirm {
       String dialogSubmitActionPath = DefaultPath.dialogSubmitAction,
       String dialogCancelTextPath = DefaultPath.dialogCancelText,
       String dialogCancelActionPath = DefaultPath.dialogCancelAction,
-      String defaultSubmitText = "Yes",
-      String defaultCacnelText = "No",
-      String defaultTitle,
-      String defaultText,
-      VoidAction defaultSubmitAction,
-      VoidAction defaultCancelAction,
+      String submitText = "Yes",
+      String cacnelText = "No",
+      String title,
+      String text,
+      VoidAction submitAction,
+      VoidAction cancelAction,
       bool popOnPress = true,
       bool willShowRepetition = false}) async {
     if (context == null) return;
-    String title = context.read(dialogTitlePath, defaultValue: defaultTitle);
-    String text = context.read(dialogTextPath, defaultValue: defaultText);
-    if (title == null || text == null) return;
+    String _title = context.read(dialogTitlePath, defaultValue: title);
+    String _text = context.read(dialogTextPath, defaultValue: text);
+    if (_title == null || _text == null) return;
     bool clicked = false;
     OverlayState overlay = context.navigator.overlay;
     do {
@@ -55,12 +55,12 @@ class UIConfirm {
           barrierDismissible: false,
           builder: (context) {
             return AlertDialog(
-              title: Text(title),
-              content: Text(text),
+              title: Text(_title),
+              content: Text(_text),
               actions: <Widget>[
                 FlatButton(
                   child: Text(context.read(dialogCancelTextPath,
-                      defaultValue: defaultCacnelText)),
+                      defaultValue: cacnelText)),
                   onPressed: () {
                     PathMap.removeAllPath([
                       dialogTitlePath,
@@ -74,13 +74,13 @@ class UIConfirm {
                     if (popOnPress)
                       Navigator.of(context, rootNavigator: true).pop();
                     context.readAction(dialogCancelActionPath,
-                        defaultAction: defaultCancelAction)();
+                        defaultAction: cancelAction)();
                     clicked = true;
                   },
                 ),
                 FlatButton(
                   child: Text(context.read(dialogSubmitTextPath,
-                      defaultValue: defaultSubmitText)),
+                      defaultValue: submitText)),
                   onPressed: () {
                     PathMap.removeAllPath([
                       dialogTitlePath,
@@ -94,7 +94,7 @@ class UIConfirm {
                     if (popOnPress)
                       Navigator.of(context, rootNavigator: true).pop();
                     context.readAction(dialogSubmitActionPath,
-                        defaultAction: defaultSubmitAction)();
+                        defaultAction: submitAction)();
                     clicked = true;
                   },
                 )